Fixed serious bug in game logic
Bug: sometimes the card table would be populated with duplicate
cards, sometimes leading to an impossible round (all cards on the
table exist within the player deck).
This was due to type mismatch between cards in the pokedex being
stored as numbers and cards in the player deck being stored as strings.
Cards in player deck were stored as strings due to taking the id from the
card HTML element.
Game logic has up to 9 player cards being replaced into the current table,
with 1 new card from the total deck. Sometimes the new card would be an integer
representation of one of the 9 replaced player cards held as strings. This results
in an impossible round.
Solution, explicitly cast id to number when inputting into the player deck.

TODO: winning game check (all cards from generation are in player deck)